requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-huggingface==0.0.3
openai==1.40.0
chromadb==0.5.3
pymupdf==1.23.8
camelot-py==0.11.0
opencv-python-headless==4.8.1.78
# NOTE: ghostscript must be installed as a system package:
#   macOS:  brew install ghostscript
#   Ubuntu: apt-get install ghostscript
#   Windows: https://www.ghostscript.com/download/gsdnld.html
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
sentence-transformers==3.1.1
